change wire data type to int for steps

+++ ios/ActivitySources/Sources/Activities/HealthKit/Entities/HKDailyMetricDataPoint.swift
@@ -2,6 +2,6 @@ import Foundation
 
 struct HKDailyMetricDataPoint: Codable {
     var date: Date
-    var steps: Double?
-    var restingHeartRate: Double?
+    var steps: Int?
+    var restingHeartRate: Int?
 }
